requirements.txt
sqlalchemy
fastapi
bcrypt
httpx
psycopg2-binary
python-dotenv
uvicorn
asyncpg
python-multipart
passlib
pydantic
pyjwt
pytest-asyncio